Add class so it can be overwritten from outside
Strangely, we have to add a class, so that a
parent component can generate enough specificity.
At least, if you want to overwrite only tags.
Before this commit, it was not possible to
overwrite styles for the <th> element:
"div :global(table th)" did not work, because
of https://github.com/sveltejs/svelte/issues/4374

Now it is possible by targeting
"div :global(table.et th)".